refactor: extract CLI commands from index.ts into focused modules
Split the 1,196-line index.ts into focused modules:
- cli/ipc-client.ts: sendOneMessage IPC utility
- cli/core-commands.ts: daemon, dev, sessions, audit, doctor, completions
- cli/config-commands.ts: config, keys, trust, memory commands
- index.ts: slim 70-line entry point with argument parsing and dispatch
